test: split the output contract by what it protects
`output_contract.rs` had grown to within a dozen lines of the hard file
limit, and the tests inside it had stopped answering one question. It
now keeps the *shape* of a command's output — headers, JSON keys,
styling — while `reporting_contract.rs` holds whether a command reports
what it did at all, which is what #1248 was about.

The fixtures move to a `harness` module both declare. Each integration
test file is its own crate, so sharing has to be a module rather than an
import.

Refs #1248
